refactor(codegen): move CLI utils to template-copy pattern
Move CLI utils code from a string template in utils-generator.ts to a
real TypeScript file at templates/cli-utils.ts, following the same
pattern as ORM templates (orm-client.ts, query-builder.ts, etc.).

Benefits:
- Type-checked at build time (was just a string before)
- IDE support (syntax highlighting, autocomplete, refactoring)
- Testable directly as a real .ts module
- Easier to maintain (no escaping backticks, no string concatenation)
- Consistent with existing template-copy pattern in the codebase
